requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.3.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typer==0.18.0
typing-inspection==0.4.1
typing_extensions==4.15.0
openai>=1.0.0

